Make network-messagepack-rpc available in tests without IO
